requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
requests==2.28.1
scikit_learn==1.0.2
streamlit==1.3.1
protobuf==3.20